feat: 상표 분석 결과 PDF 리포트 실시간 생성 기능 구현

+++ backend/src/main/java/com/royalty/backend/mypage/dto/BrandHistoryDTO.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 public class BrandHistoryDTO {
     private String version;          // v1, v2 ...
     private String imagePath;        // 당시 로고
-    private Double imageSimilarity;  // 당시 이미지 유사도
-    private Double textSimilarity;   // 당시 텍스트 유사도
+    private Float imageSimilarity; // 이미지 유사도 (%)
+    private Float textSimilarity;  // 텍스트 유사도 (%)
     private LocalDateTime createdAt; // 변경일
 }
